Separate out the factory from the Product
This keeps things more separate and also allows us to avoid some
horrible circular import type issues coming up.
