+++ lms/product/__init__.py
@@ -1,7 +1,8 @@
+from lms.product.factory import get_product_from_request
 from lms.product.product import Product
 
 
 def includeme(config):
     config.add_request_method(
-        Product.from_request, name="product", property=True, reify=True
+        get_product_from_request, name="product", property=True, reify=True
     )
