+++ tests/unit/lms/product/product_test.py
@@ -1,36 +1,11 @@
-import pytest
+from unittest.mock import sentinel
 
 from lms.product.product import Product
 
-PRODUCT_NAMES = [
-    ("BlackbaudK12", Product.Family.BLACKBAUD),
-    ("BlackboardLearn", Product.Family.BLACKBOARD),
-    ("canvas", Product.Family.CANVAS),
-    ("desire2learn", Product.Family.D2L),
-    ("moodle", Product.Family.MOODLE),
-    ("schoology", Product.Family.SCHOOLOGY),
-    ("sakai", Product.Family.SAKAI),
-    ("wut", Product.Family.UNKNOWN),
-    ("", Product.Family.UNKNOWN),
-    (None, Product.Family.UNKNOWN),
-]
-
 
 class TestProduct:
-    @pytest.mark.parametrize("value,expected", PRODUCT_NAMES)
-    def test_from_request_exact_match_lti(self, value, expected, pyramid_request):
-        pyramid_request.lti_params["tool_consumer_info_product_family_code"] = value
+    def test_from_request(self):
+        product = Product.from_request(sentinel.request)
 
-        assert Product.from_request(pyramid_request).family == expected
-
-    @pytest.mark.parametrize("value,expected", PRODUCT_NAMES)
-    def test_from_request_api(self, value, expected, pyramid_request):
-        pyramid_request.content_type = "application/json"
-        pyramid_request.json = {"lms": {"product": value}}
-
-        assert Product.from_request(pyramid_request).family == expected
-
-    def test_from_request_canvas_custom(self, pyramid_request):
-        pyramid_request.lti_params = {"custom_canvas_course_id": "course_id"}
-
-        assert Product.from_request(pyramid_request).family == Product.Family.CANVAS
+        assert isinstance(product, Product)
+        assert product.family == Product.Family.UNKNOWN
